Update gengui.py
Adding write_only_key (https://pysimplegui.readthedocs.io/en/latest/#write_only_key-modifier) to a few outputs to... optimize? I dunno man

Also did a quick and dirty change to fix a bug where it would crash if you tried to generate a new first name, but there was no space character (" ") in the name output box. Tried to make it read-only, but apparently that doesn't exist... which makes sense because it's an input box, but still